backend/pyproject.toml
[tool.poetry]
name = "mykeymanager-backend"
version = "0.1.0"
description = "Backend API for MyKeyManager license management"
authors = ["Owner <info@acwild.it>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras=["standard"], version="^0.30.0"}
python-multipart = "^0.0.9"
email-validator = "^2.1.0.post1"
passlib = {extras=["bcrypt"], version="^1.7.4"}
pyjwt = "^2.8.0"
SQLAlchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
slowapi = "^0.1.9"
itsdangerous = "^2.2.0"
redis = "^5.0.4"
cryptography = "^42.0.8"
pydantic-settings = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
black = "^24.4.2"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
